% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_add_projection_and_write.R
\name{add_and_write_projection}
\alias{add_and_write_projection}
\alias{add_and_write_pca}
\alias{add_and_write_sma}
\alias{add_and_write_pls}
\alias{add_and_write_lda}
\alias{add_and_print_pca}
\title{Run pca and print results}
\usage{
add_and_write_projection(object, method = "pca", implementation = NULL,
  result_dir, feature_plots = autonomics.plot::default_feature_plots(object),
  x = autonomics.plot::default_x(object, feature_plots),
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object, color_var),
  shape_var = autonomics.plot::default_shape_var(object),
  group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object),
  line = autonomics.plot::default_line(object), na.impute = FALSE)

add_and_write_pca(object, ...)

add_and_write_sma(object, ...)

add_and_write_pls(object, ...)

add_and_write_lda(object, ...)

add_and_print_pca(...)
}
\arguments{
\item{object}{eset}

\item{method}{'pca', 'pls', or 'lda'}

\item{implementation}{character}

\item{result_dir}{result dir}

\item{feature_plots}{subset of \code{\link[autonomics.plot]{FEATURE_PLOTS}}}

\item{x}{svar mapped to x in feature plots}

\item{color_var}{svar mapped to color in feature plots}

\item{color_values}{color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape in feature plots}

\item{group_var}{svar mapped to group in feature plots}

\item{txt_var}{svar mapped to txt   in feature plots}

\item{line}{whether to connect points in feature plots with a line (logical)}

\item{na.impute}{whether to impute missing values}

\item{...}{for backward compatibility}
}
\value{
eset with pca results
}
\description{
Run pca and print results
}
\examples{
require(magrittr)
result_dir <- tempdir() \%T>\% message()
if (require(autonomics.data)){
   object <- autonomics.data::billing2016 
   object \%>\% autonomics.explore::add_and_write_pca(result_dir = result_dir)
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% add_and_write_pca(result_dir = result_dir, color_var = 'condition')
}
if (require(billing.differentiation.data)){
   require(magrittr)
   result_dir <- tempdir() \%T>\% message()
   billing.differentiation.data::rna.voomcounts \%>\% 
      autonomics.explore::add_and_write_pca(result_dir = result_dir)
}
if (require(halama.2016)){
   result_dir <- tempdir() \%T>\% message()
   halama.2016::cell.metabolites \%>\% autonomics.explore::add_and_write_pca(
      result_dir   = result_dir,
      color_var    = 'GROUP_DESCRIPTION', 
      color_values = c(Control = 'orange', Vehicle = 'red', `Concentration 1` = 'green', 
                      `Concentration 2` = 'blue'))
}
}
