% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_add_pca_to_eset.R
\name{add_projection_to_eset}
\alias{add_projection_to_eset}
\alias{add_pca_to_eset}
\alias{add_sma_to_eset}
\alias{add_lda_to_eset}
\alias{add_pls_to_eset}
\title{Add PCA/LDA/PLS/SMA results to eset}
\usage{
add_projection_to_eset(object, method = c("pca", "lda", "pls")[1],
  implementation = NULL, na.impute = FALSE, ndim = 2)

add_pca_to_eset(object, ...)

add_sma_to_eset(object, ...)

add_lda_to_eset(object, ...)

add_pls_to_eset(object, ...)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{method}{'pca' or 'lda'}

\item{implementation}{(character)}

\item{na.impute}{TRUE or FALSE}

\item{ndim}{number of dimensions to include}

\item{...}{passed to add_projection_to_eset}
}
\description{
Add PCA/LDA/PLS/SMA results to eset
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% autonomics.explore::add_pca_to_eset()
   subramanian.2016::metabolon \%>\% autonomics.explore::add_lda_to_eset()
   subramanian.2016::metabolon \%>\% autonomics.explore::add_pls_to_eset()
}

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% autonomics.explore::add_pca_to_eset(ndim = 3)
   object \%<>\% autonomics.explore::add_pca_to_eset()
   object \%>\% autonomics.import::svars()
   object \%>\% autonomics.import::fvars()
}

# STEM CELL DIFFERENTIATION
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% autonomics.explore::add_pca_to_eset()
}

if (require(atkin.2014)){
   object <- atkin.2014::soma 
   object \%>\% autonomics.explore::add_pca_to_eset()
}
}
