% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plot_projection_features.R
\name{extract_top_and_bottom}
\alias{extract_top_and_bottom}
\title{Extract top pca features}
\usage{
extract_top_and_bottom(object, dim = 1, n = 16, na.impute)
}
\arguments{
\item{object}{eset object}

\item{dim}{principal component (integer)}

\item{n}{number of top features to be selected}

\item{na.impute}{logical}
}
\description{
Extract top pca features
}
