% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_projected_samples_and_features.R
\name{layout_sample_projections_and_features}
\alias{layout_sample_projections_and_features}
\title{Create graphics layout for mpa feature + sample plot}
\usage{
layout_sample_projections_and_features(feature_plot)
}
\arguments{
\item{feature_plot}{feature plot (character)}
}
\value{
graphics layout matrix
}
\description{
Create graphics layout for mpa feature + sample plot
}
