% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{pca}
\alias{pca}
\title{Principal Component Analysis (PCA)}
\usage{
pca(object, dims = 1:2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{dims}{dimensions vector}

\item{...}{included only to keep \code{\link{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  samples x dims
     features =  feature loadings,     # matrix: features x dims
     var      = variance percentages   # vector:        1 x dims
}
}
\description{
Returns \code{PCA} sample scores, feature loadings, and component variances.
}
\details{
\code{PCA} is performed using the (\code{NA}-friendly) \code{NIPALS} algorithm \cr
instead of the mathematically elegant, but \code{NA}-unfriendly \code{SVD} formulation. \cr\preformatted{SVD:       X = U*D*V' (U = L singular vectors, V = R singular vectors, D = singular values)
      Scores = U*D    (or U*sqrt(D)   in SMA)
    Loadings = V      (or   sqrt(D)*V in SMA)

NIPALS:    X = T*P'   (no further decomposition of scores, so SMA not possible)
      Scores = T = U*D
    Loadings = P = V
}
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% autonomics.explore::pca() \%>\% str()
}
if (require(halama.2016)){
   object <- halama.2016::cell.metabolites 
   object \%>\% autonomics.explore::pca() \%>\% str()
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.explore::pca()  \%>\%  str()
}
}
\seealso{
\code{\link[autonomics.explore]{sma}},
\code{\link[autonomics.explore]{lda}},
\code{\link[autonomics.explore]{pls}}
}
\author{
Aditya Bhagwat
}
