% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot_pca_samples.R
\name{plot_projected_samples}
\alias{plot_projected_samples}
\alias{plot_pca_samples}
\alias{plot_sma_samples}
\alias{plot_lda_samples}
\alias{plot_pls_samples}
\title{Plot PCA/LDA/PLS sample scores}
\usage{
plot_projected_samples(object, ..., listed_objects = NULL, method = c("pca",
  "lda", "sma", "pls")[1], implementation = c("mixOmics::plsda",
  "mixOmics::splsda", "ropls::opls")[1], dims = 1:2, color_var = "subgroup",
  color_values = autonomics.plot::default_color_values(object, color_var),
  shape_var = autonomics.plot::default_shape_var(object), size_var = NULL,
  txt_var = autonomics.plot::default_txt_var(object), group_var = NULL,
  split_var = NULL, facet_var = NULL, scales = ifelse(is.null(facet_var),
  "fixed", "free"), labs = list(color =
  stringi::stri_trans_totitle(color_var), shape =
  stringi::stri_trans_totitle(shape_var)), nrow = NULL,
  mention_method = ifelse(is.null(facet_var), TRUE, FALSE), title = NULL,
  na.impute = FALSE, legend.position = "right")

plot_pca_samples(object, ...)

plot_sma_samples(object, ...)

plot_lda_samples(object, ...)

plot_pls_samples(object, ...)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or Elist}

\item{...}{Further SummarizedExperiment(s), eSet(s), or Elist(s)}

\item{listed_objects}{Further SummarizedExperiment(s), eSet(s), or Elist(s), wrapped in a \code{link{list}}}

\item{method}{'pca', 'lda', 'pls', 'sma'}

\item{implementation}{which implementation of pls to use (see \code{\link{project}})}

\item{dims}{dimensions}

\item{color_var}{svar mapped to color}

\item{color_values}{color values vector}

\item{shape_var}{svar mapped to shape}

\item{size_var}{svar mapped to size}

\item{txt_var}{svar mapped to txt}

\item{group_var}{svar mapped to group}

\item{split_var}{svar on which to split data prior to transformation}

\item{facet_var}{svar on which to facet sample biplot}

\item{scales}{'free' or 'fixed'}

\item{labs}{named \code{\link{list}} handed on to \code{\link[ggplot2]{labs}}}

\item{nrow}{integer}

\item{mention_method}{Whether to use \code{method} in title and/or facet labels}

\item{title}{title}

\item{na.impute}{TRUE or FALSE}

\item{legend.position}{character}
}
\description{
Plot PCA/LDA/PLS sample scores
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.explore::plot_pca_samples()
   object \%>\% autonomics.explore::plot_pca_samples(dims = c(3,4))
   
   object \%>\% autonomics.explore::plot_projected_samples(
       method = c('pca', 'lda', 'sma', 'pls'),
       facet_var = c('OK PCA', 'Not so nice LDA', 'Nice SMA', 'Very nice PLS'),
       nrow = 2)
}
}
\author{
Aditya Bhagwat, Johannes Graumann
}
