% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_pca_samples_and_features.R
\name{plot_projected_samples_and_features}
\alias{plot_projected_samples_and_features}
\alias{plot_pca_samples_and_features}
\alias{plot_sma_samples_and_features}
\alias{plot_lda_samples_and_features}
\alias{plot_pls_samples_and_features}
\title{Plot pca samples and features}
\usage{
plot_projected_samples_and_features(object, method, implementation = NULL,
  dims = c(1, 2), na.impute = FALSE,
  feature_plot = autonomics.plot::default_feature_plots(object)[1],
  result_dir = NULL, x = autonomics.plot::default_x(object, feature_plot),
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object, color_var),
  shape_var = NULL, group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object),
  line = autonomics.plot::default_line(object), n = 9)

plot_pca_samples_and_features(object, ...)

plot_sma_samples_and_features(object, ...)

plot_lda_samples_and_features(object, ...)

plot_pls_samples_and_features(object, ...)
}
\arguments{
\item{object}{eset}

\item{method}{'pca', 'lda', or 'pls'}

\item{implementation}{NULL or "package::function"}

\item{dims}{pc dimensions to plot}

\item{na.impute}{TRUE or FALSE}

\item{feature_plot}{value in \code{\link[autonomics.plot]{FEATURE_PLOTS}}}

\item{result_dir}{NULL or result directory path}

\item{x}{svar mapped to x in feature plots}

\item{color_var}{svar mapped to color in sample and feature plots}

\item{color_values}{color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape in sample and feature plots}

\item{group_var}{svar mapped to group in feature plots}

\item{txt_var}{svar mapped to txt in feature plots}

\item{line}{whether to connect points in feature plot with line (logical)}

\item{n}{no of features to be plotted}

\item{...}{passed to plot_projected_samples_and_features}
}
\description{
Plots pca sample and features results in a combined plot
}
\examples{
require(magrittr)
result_dir <- tempdir() \%T>\% message()
\dontrun{

     # STEM CELL COMPARISON
     if (require(autonomics.data)){
        object <- autonomics.data::stemcomp.proteinratios
        object \%>\% autonomics.explore::plot_pca_samples_and_features()
        object \%>\% autonomics.explore::plot_pca_samples_and_features(
                      feature_plot = 'distributions')
        object \%>\% autonomics.explore::plot_pca_samples_and_features(
                      feature_plot = 'distributions', na.impute = TRUE)
        object \%>\% autonomics.explore::plot_pca_samples_and_features(
                      result_dir = result_dir)
        object \%>\% autonomics.explore::plot_pca_samples_and_features(
                      result_dir = result_dir, feature_plot = 'bars')
     }
     
     # STEM CELL DIFFERENTIATION
     if (require(billing.differentiation.data)){
        object <- billing.differentiation.data::rna.voomcounts
        object \%>\% autonomics.explore::plot_pca_samples_and_features()
     }
     
     # GLUTAMINASE
     if (require(autonomics.data)){
        autonomics.data::glutaminase \%>\% autonomics.explore::plot_pca_samples_and_features(
           color_var    = 'GROUP_DESCRIPTION',
           color_values = c(Control = 'orange', Vehicle = 'red', `Concentration 1` = 'green',
                           `Concentration 2` = 'blue'))
     }
     
     if (require(subramanian.2016)){
        object <- subramanian.2016::metabolon
        object \%>\% autonomics.explore::plot_pca_samples_and_features(n = 4)
        object \%>\% autonomics.explore::plot_lda_samples_and_features(n = 4)
     
        object \%>\% autonomics.explore::plot_pca_samples_and_features(
                      color_var = 'condition', feature_plot = 'boxes')
     }
}
}
