% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_plot_sample_distributions_and_projection.R
\name{plot_sample_distrs_n_pca}
\alias{plot_sample_distrs_n_pca}
\title{Plot sample distributions and pca}
\usage{
plot_sample_distrs_n_pca(object, descr = "",
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object),
  shape_var = autonomics.plot::default_shape_var(object),
  txt_var = autonomics.plot::default_txt_var(object),
  displayed_features = NULL)
}
\arguments{
\item{object}{eset}

\item{descr}{description: used as plot title and plot name}

\item{color_var}{svar mapped to color}

\item{color_values}{color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape}

\item{txt_var}{svar mapped to txt}

\item{displayed_features}{features to be displayed in the sample distributions (vector of numeric indexes or character \code{feature_id}s)}
}
\value{
list(descr.distr = plot1, descr.pca = plot2)
}
\description{
Plot sample distributions and pca
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
      autonomics.explore::plot_sample_distrs_n_pca('rna.voomcounts')
}

# GLUTAMINASE
if (require(autonomics.data)){
   autonomics.data::glutaminase \%>\% autonomics.explore::plot_sample_distrs_n_pca(
      descr = 'custom colors',
      color_var    = 'GROUP_DESCRIPTION',
      color_values = c(Control = 'orange', Vehicle = 'red', `Concentration 1` = 'green',
                      `Concentration 2` = 'blue'))
}

if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% 
   autonomics.explore::plot_sample_distrs_n_pca()
}
}
