% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{pls}
\alias{pls}
\title{Partial least squares analysis (PLS)}
\usage{
pls(object, implementation = NULL, ndim = 2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{implementation}{'mixOmics::plsda', 'mixOmics::splsda', or 'ropls::opls'}

\item{ndim}{Number of latent variables to include in results}

\item{...}{only inlcuded to keep \code{\link[autonomics.explore]{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  nsample x ndim
     features =  feature loadings,     # matrix: nfeature x ndim
     var      = variance percentages   # vector:        1 x ndim
}
}
\description{
Returns \code{PLS} sample scores, feature loadings, and latent variable variances.
}
\details{
\code{PLS} is performed using the (\code{NA}-friendly) \code{NIPALS} algorithm.
}
\examples{
require(magrittr)

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   object \%>\% autonomics.explore::pls() \%>\% str()
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.explore::pls() \%>\% str()
   \dontrun{ # slow
      object \%>\% autonomics.explore::pls(implementation = 'ropls::opls'     ) \%>\% str()
   }
}
}
\seealso{
\code{\link[autonomics.explore]{pca}},
\code{\link[autonomics.explore]{sma}},
\code{\link[autonomics.explore]{lda}}
}
\author{
Aditya Bhagwat
}
