% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_project.R
\name{pca}
\alias{pca}
\alias{sma}
\alias{lda}
\alias{pls}
\alias{project}
\title{PCA, SMA, LDA, or PLS transform}
\usage{
pca(object, na.impute = FALSE, dims = 1:2, ...)

sma(object, na.impute = FALSE, dims = 1:2, ...)

lda(object, na.impute = FALSE, dims = 1:2, ...)

pls(object, na.impute = FALSE, implementation = NULL, dims = 1:2, ...)

project(object, method = "pca", na.impute = FALSE, implementation = NULL,
  dims = 1:2, dims_conflict = "warn")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{na.impute}{TRUE or FALSE}

\item{dims}{\code{\link{numeric}} vector indicating what 'dimentions' (e.g. PCA: dimensions == components) to return}

\item{...}{further parameters}

\item{implementation}{which implementation of the method to use}

\item{method}{'pca', 'sma', 'lda', 'pls'}

\item{dims_conflict}{how to react to a conflict between what \code{lda} can deliver (\code{dims = 1:2}) and what dims are requested}
}
\value{
list(samples, features, var)
}
\description{
Perform principal component analysis (PCA), spectral map analysis (SMA), 
linear discriminant analysis (LDA), or partial least squares analysis (PLS).
}
\details{
Compute SMA and LDA using SVD (utilizable for all methods, but restricted to non NA features).
Compute PCA and PLS using NIPALS (utilizable for PCA and PLS only, but allows features with NAS).

PCA and SMA are related: their SVD is identical, but score and loading computation differs 
(Wouters et al. 2003 - see references) \cr
   PCA Scores: U * D                              \cr
   SMA Scores: U * D^(1/2)                        \cr 
                                                  \cr
   PCA Loadings: V                                \cr
   SMA Loadings: D^(1/2)*V                        \cr
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.explore::pca() \%>\% str()
   object \%>\% autonomics.explore::pca(dims = 1:6) \%>\% str()
   object \%>\% autonomics.explore::project(method = 'lda') \%>\% str()
\dontrun{
   # Fails, as max dim length(subgroup) - 1
   object \%>\%
     autonomics.explore::project(method = 'lda', dims = 1:3) \%>\%
     str()}
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.explore::pca()  \%>\%  str()
   object \%>\% autonomics.explore::sma()  \%>\%  str()
   
   object \%>\% autonomics.explore::lda()  \%>\%  str()
   object \%>\% autonomics.import::filter_samples(
                 subgroup \%in\% autonomics.import::subgroup_levels(.)[1:2]) \%>\% 
              autonomics.explore::lda()
              
   object \%>\% autonomics.explore::pls()  \%>\%  str()
   object \%>\% autonomics.explore::pls(implementation = 'mixOmics::splsda') \%>\% str()
   \dontrun{object \%>\% autonomics.explore::pls(implementation = 'ropls::opls'     ) \%>\% str()}
}
if (require(halama.2016)){
   object <- halama.2016::cell.metabolites
   object \%>\% autonomics.explore::project('lda') \%>\% str()
   object \%>\% autonomics.explore::project('pca') \%>\% str()
}
}
\references{
PCA 
SMA: Wouters et al (2003) Graphical exploration of gene expression data: a comparative study of 
     three multivariate methods
}
