% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{sma}
\alias{sma}
\title{Spectral Map Analysis (SMA)}
\usage{
sma(object, na.impute = FALSE, dims = 1:2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{na.impute}{whether to \code{\link[autonomics.preprocess]{impute}} prior to \code{SMA}}

\item{dims}{dimensions vector}

\item{...}{used to keep \code{\link{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  samples x dims
     features =  feature loadings,     # matrix: features x dims
     var      = variance percentages   # vector:        1 x dims
}
}
\description{
Returns \code{SMA} sample scores, feature loadings, and component variances.
}
\details{
\code{SMA} is based on the (\code{NA}-unfriendly) \code{SVD}, so only \code{NA}-free features are used. \cr
\code{SMA} and \code{PCA} are related: their \code{SVD} is identical, but score and loading computation differ:\preformatted{   PCA Scores  : U * D
   SMA Scores  : U * D^(1/2)
   
   PCA Loadings:         V
   SMA Loadings: D^(1/2)*V
}
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% autonomics.explore::sma() \%>\% str()
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.explore::sma()  \%>\%  str()
}
}
\references{
Wouters et al (2003) Graphical exploration of gene expression data: a comparative study of
three multivariate methods.
}
\seealso{
\code{\link[autonomics.explore]{pca}},
\code{\link[autonomics.explore]{lda}},
\code{\link[autonomics.explore]{pls}}
}
\author{
Aditya Bhagwat, Laure Cougnaud
}
