% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_add_projection_and_write.R
\name{write_projected_features}
\alias{write_projected_features}
\title{Write pca features}
\usage{
write_projected_features(object, result_dir, method = "pca")
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{result_dir}{result dir}

\item{method}{which character}
}
\description{
Write pca features
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   result_dir <- tempdir() \%T>\% message()
   object \%>\% autonomics.explore::add_pca_to_eset() \%>\% 
              autonomics.explore::write_projected_features(result_dir = result_dir)
}
}
