% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{lda}
\alias{lda}
\title{Linear discriminant analysis (LDA)}
\usage{
lda(object, na.impute = FALSE, ndim = 2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{na.impute}{currently ignored. May be used in future.}

\item{ndim}{Number of linear discriminants to include in results}

\item{...}{only included to keep \code{\link[autonomics.explore]{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  nsample x ndim
     features =  feature loadings,     # matrix: nfeature x ndim
     var      = variance percentages   # vector:        1 x ndim
}
}
\description{
Returns \code{LDA} sample scores, feature loadings, and discriminant variances.
}
\details{
\code{LDA} relies on the (\code{NA}-unfriendly) \code{SVD}, so only \code{NA}-free features are included in the analysis.
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.explore::lda() \%>\% str()
   \dontrun{ # Fails, as max dim length(subgroup) - 1
      object \%>\% autonomics.explore::lda(ndim = 3) \%>\% str()
   }
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% autonomics.explore::lda() \%>\% str()
}
if (require(halama.2016)){
   object <- halama.2016::cell.metabolites
   object \%>\% autonomics.explore::lda() \%>\% str()
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.explore::lda()  \%>\%  str()
}
}
\seealso{
\code{\link[autonomics.explore]{pca}},
\code{\link[autonomics.explore]{sma}},
\code{\link[autonomics.explore]{pls}}
}
\author{
Aditya Bhagwat, Laure Cougnaud
}
