% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot_pca_samples_version1.R
\name{make_projected_samples_df}
\alias{make_projected_samples_df}
\title{Make datatable to plot PCA, LDA, or PLS sample scores}
\usage{
make_projected_samples_df(projection, object, dims, color_var, shape_var,
  size_var, txt_var, split_var, facet_var, group_var)
}
\arguments{
\item{projection}{output from pca, lda, or pls}

\item{object}{SummarizedExperiment, eSet, or eList}

\item{dims}{PCA/LDA dimensions}

\item{color_var}{svar mapped to color}

\item{shape_var}{svar mapped to shape}

\item{size_var}{svar mapped to size}

\item{txt_var}{svar mapped to txt}

\item{split_var}{svar on which object is splitted prior to lda analysis}

\item{facet_var}{svar on which plot is faceted}

\item{group_var}{svar mapped to group_var}
}
\description{
Make datatable to plot PCA, LDA, or PLS sample scores
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   
   autonomics.explore::pca(object) \%>\% 
   autonomics.explore::make_projected_samples_df(
      object, dims = c(1,2), color_var = 'subgroup', 
      shape_var = NULL, size_var = NULL, txt_var = NULL, 
      split_var = NULL, facet_var = NULL, group_var = NULL) \%>\% 
   print()
   
   autonomics.explore::pca(object, ndim=4) \%>\% 
   autonomics.explore::make_projected_samples_df(
      object, dims = 3:4, color_var = 'subgroup', 
      shape_var = NULL, size_var = NULL, txt_var = NULL, 
      split_var = NULL, facet_var = NULL, group_var = NULL) \%>\% 
   print()
}
}
