% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{pca}
\alias{pca}
\title{Principal Component Analysis (PCA)}
\usage{
pca(object, ndim = 2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ndim}{Number of PCA components to include in results}

\item{...}{included only to keep \code{\link{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  nsamples x ndim
     features =  feature loadings,     # matrix: nfeatures x ndim
     var      = variance percentages   # vector:         1 x ndim
}
}
\description{
Returns \code{PCA} sample scores, feature loadings, and component variances.
}
\details{
\code{PCA} is performed using the (\code{NA}-friendly) \code{NIPALS} algorithm \cr
instead of the mathematically elegant, but \code{NA}-unfriendly \code{SVD} formulation. \cr\preformatted{SVD:       X = U*D*V' (U = L singular vectors, V = R singular vectors, D = singular values)
      Scores = U*D    (or U*sqrt(D)   in SMA)
    Loadings = V      (or   sqrt(D)*V in SMA)

NIPALS:    X = T*P'   (no further decomposition of scores, so SMA not possible)
      Scores = T = U*D
    Loadings = P = V
}
}
\examples{
require(magrittr)

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   object \%>\% autonomics.explore::pca() \%>\% str()
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.explore::pca() \%>\% str()
}

}
\seealso{
\code{\link[autonomics.explore]{sma}},
\code{\link[autonomics.explore]{lda}},
\code{\link[autonomics.explore]{pls}}
}
\author{
Aditya Bhagwat
}
