% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plot_pca_features.R
\name{plot_projection_features}
\alias{plot_projection_features}
\alias{plot_pca_features}
\alias{plot_sma_features}
\alias{plot_lda_features}
\alias{plot_pls_features}
\title{Plot pca features}
\usage{
plot_projection_features(object, method, implementation = NULL,
  feature_plot = autonomics.plot::default_feature_plots(object)[1],
  x = autonomics.plot::default_x(object, feature_plot),
  color_var = autonomics.plot::default_color_var(object),
  color_values = autonomics.plot::default_color_values(object, color_var),
  shape_var = autonomics.plot::default_shape_var(object),
  group_var = autonomics.plot::default_group_var(object),
  txt_var = autonomics.plot::default_txt_var(object), facet_var = NULL,
  line = autonomics.plot::default_line(object),
  fvars = autonomics.plot::default_fvars(object), dim = 1, n = 9,
  na.impute = FALSE, title = sprintf("X\%d", dim), file = NULL,
  legend.position = "right")

plot_pca_features(object, ...)

plot_sma_features(object, ...)

plot_lda_features(object, ...)

plot_pls_features(object, ...)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or Elist}

\item{method}{'pca' or 'lda'}

\item{implementation}{'character' or NULL}

\item{feature_plot}{value in \code{\link[autonomics.plot]{FEATURE_PLOTS}}}

\item{x}{svar mapped to x}

\item{color_var}{svar mapped to color}

\item{color_values}{color vector (names = color_var levels, values = colors)}

\item{shape_var}{svar mapped to shape}

\item{group_var}{svar mapped to group}

\item{txt_var}{svar mapped to txt}

\item{facet_var}{svar used for faceting}

\item{line}{whether to connect points with line (logical)}

\item{fvars}{fvars used for plot annotation}

\item{dim}{principal component dimension}

\item{n}{number of top features to plot}

\item{na.impute}{TRUE or FALSE}

\item{title}{title}

\item{file}{file}

\item{legend.position}{ggplot theme argument}

\item{...}{passed to plot_projection_features}
}
\description{
Plots top pca features.
Uses pca factor loadings in fdata when available.
When not, add_pca_to_eset is run prior to plotting
}
\examples{
require(magrittr)
# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% autonomics.explore::plot_pca_features(n=9)
   object \%>\% autonomics.explore::plot_pca_features(feature_plot = 'bars')
}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   object \%>\% autonomics.explore::plot_pca_features(n=9)
   object \%>\% autonomics.explore::plot_pca_features(dim=2, n=9)
   object \%>\% autonomics.explore::plot_pca_features()
}

# GLUTAMINASE
if (require(autonomics.data)){
   autonomics.data::glutaminase \%>\% autonomics.explore::plot_pca_features(
      color_var    = 'GROUP_DESCRIPTION', 
      color_values = c(Control = 'orange', Vehicle = 'red', `Concentration 1` = 'green', 
                      `Concentration 2` = 'blue'))
}

if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.explore::plot_pca_features(feature_plot = 'boxes')
   object \%>\% autonomics.explore::plot_pls_features(feature_plot = 'boxes')
   object \%>\% autonomics.explore::plot_lda_features(feature_plot = 'boxes')
}

if (require(atkin.2014)){
   object <- atkin.2014::soma
   object \%>\% autonomics.explore::plot_pca_features()
}

}
