% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{pls}
\alias{pls}
\title{Partial least squares analysis (PLS)}
\usage{
pls(object, implementation = NULL, dims = 1:2, ...)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{implementation}{'mixOmics::plsda', 'mixOmics::splsda', or 'ropls::opls'}

\item{dims}{dimensions vector}

\item{...}{only inlcuded to keep \code{\link[autonomics.explore]{project}} generic}
}
\value{
\preformatted{list(samples  =   sample scores,       # matrix:  samples x dims
     features =  feature loadings,     # matrix: features x dims
     var      = variance percentages   # vector:        1 x dims
}
}
\description{
Returns \code{PLS} sample scores, feature loadings, and latent variable variances.
}
\details{
\code{PLS} is performed using the (\code{NA}-friendly) \code{NIPALS} algorithm.
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% autonomics.explore::pls() \%>\% str()
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.explore::pls()  \%>\%  str()
   object \%>\% autonomics.explore::pls(implementation = 'mixOmics::splsda') \%>\% str()
   \dontrun{ # slow
      object \%>\% autonomics.explore::pls(implementation = 'ropls::opls'     ) \%>\% str()
   }
}
}
\seealso{
\code{\link[autonomics.explore]{pca}},
\code{\link[autonomics.explore]{sma}},
\code{\link[autonomics.explore]{lda}}
}
\author{
Aditya Bhagwat
}
