% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pca.R
\name{project}
\alias{project}
\title{Project}
\usage{
project(object, method = "pca", na.impute = FALSE, implementation = NULL,
  ndim = 2)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{method}{'pca', 'sma', 'lda', 'pls'}

\item{na.impute}{TRUE or FALSE}

\item{implementation}{which implementation of the method to use}

\item{ndim}{Number of 'dimensions' (e.g. PCA: dimensions == components) to return}
}
\value{
list(samples, features, var)
}
\description{
Return sample scores, feature loadings, and component variances.
}
\examples{
require(magrittr)

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   object \%>\% autonomics.explore::project() \%>\% str()
}

# GLUTAMINASE
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.explore::project() \%>\% str()
}
}
\seealso{
\code{\link[autonomics.explore]{pca}}, 
         \code{\link[autonomics.explore]{sma}}, 
         \code{\link[autonomics.explore]{lda}}, 
         \code{\link[autonomics.explore]{pls}}
}
\author{
Aditya Bhagwat
}
